tg_bot_app/requirements.txt
aiogram==3.3.0
scikit-learn==1.2.2
pillow==10.2.0
